png/utils: Move conversions from core to png crate
