+++ zune-png/src/utils/sse.rs
